Media: B52ISP: improve b52 sensor hal V/H flip support
1. Improve it to support both camera sensor have different V/H
   flip regiters and have one V/H flip register.
2. Set ov13850 hflip to its original value.
3. Fix buffer VA mapping error, add data offset to the VA.

this patch is sync for kernel3.14

+++ drivers/media/platform/mmp-isp/b52-reg.c
@@ -906,7 +906,7 @@ int b52_rw_pipe_ctdata(int pipe_id, int write,
 
 	for (i = 0; i < nr_entry; i++, entry++) {
 		for (j = 0; j < tab_sz; j++) {
-			offset = entry->reg - table[j].start;
+			offset = (long)entry->reg - (long)table[j].start;
 			if ((offset >= 0) && (entry->reg < table[j].end))
 				goto addr_match;
 		}
